Adding evaluation

+++ src/types/VariantType.hpp
@@ -9,7 +9,6 @@ namespace jasl {
 
     /// Represents the set of built-in types
     using VariantType = std::variant<bool, int64_t,
-                                     double, std::string,
+                                     long double, std::string,
                                      uint8_t, std::vector<std::string>>;
-
 }
